requirements.txt
alembic>=1.14.0
cryptocode>=0.1
fastapi[standard]>=0.115.6
google-cloud-pubsub
google-cloud-storage
jwt
protobuf>=5.29.1
psycopg2
pydantic>=2.10.3
python-decouple>=3.8
redis_py_cluster>=2.1.3
spotipy>=2.24.0
sqlalchemy>=2.0.36
uvicorn>=0.32.1
